// src/store/modules/user/mutation-types.ts
/*
 * @Description: user mutations type
 * @Author: ZY
 * @Date: 2020-12-23 10:25:37
 * @LastEditors: ZY
 * @LastEditTime: 2020-12-28 11:33:53
 */

export enum UserMutationTypes {
  SET_TOKEN = "SET_TOKEN",
  SET_NAME = "SET_NAME",
  SET_AVATAR = "SET_AVATAR",
  SET_INTRODUCTION = "SET_INTRODUCTION",
  SET_ROLES = "SET_ROLES",
  SET_EMAIL = "SET_EMAIL",
  // ***
  SET_ROUTES_FROM_API = "SET_ROUTES_FROM_API",
  SET_USER_NAME = "SET_USER_NAME",
  SET_FULL_NAME = "SET_FULL_NAME",
  SET_DICTS = "SET_DICTS",
  SET_CURRENT_WAREHOUSE_INFO = "SET_CURRENT_WAREHOUSE_INFO",
  SET_WAREHOUSE_LIST = "SET_WAREHOUSE_LIST",
  SET_WAREHOUSE_PERIOD_LIST = "SET_WAREHOUSE_PERIOD_LIST",
}